% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyConstr.R
\name{myttest}
\alias{myttest}
\title{myconstructor}
\usage{
myttest(x, y, alpha, paired)
}
\arguments{
\item{x}{xvalue}

\item{y}{yvalue}

\item{alpha}{alpha value}

\item{paired}{paired value}
}
\value{
a list
}
\description{
myconstructor
}
\examples{
\dontrun {myttest(x, y,alpha, paired)}
}
