% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotf.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot Title}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{...}{the remaining arguments}
}
\value{
A plot of the data
}
\description{
The plot title creates boxplots of the data when non - paired data is analyzed.
It also will make a boxplot of differences in data when paired = TRUE
The Confidence Interval for the difference of means is plotted inside the boxplot of differences.
}
\examples{
\dontrun {plot.Rttest}
}
