% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Printf.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print Function}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
Rttest
}
\description{
This function prints to the command line  A (1-alpha)100\% confidence interval for mux-muy
}
